refactor: refactors collector services to aling with architectural definitions
